[SYRIUS-77551] number instead of long

+++ src/main/resources/db/migration/oracle/person/V1.201911210800__addLong.sql
@@ -1,1 +1,1 @@
-ALTER TABLE person ADD partner_nr LONG;
+ALTER TABLE person ADD partner_nr number(19,0);
